11/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
mutagen>=1.46.0 